+++ common/src/main/java/smartin/miapi/mixin/client/CapeFeatureRendererMixin.java
@@ -1,0 +1,29 @@
+package smartin.miapi.mixin.client;
+
+import net.minecraft.client.network.AbstractClientPlayerEntity;
+import net.minecraft.client.render.VertexConsumerProvider;
+import net.minecraft.client.render.entity.feature.CapeFeatureRenderer;
+import net.minecraft.client.util.math.MatrixStack;
+import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.item.ItemStack;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import smartin.miapi.item.modular.items.ModularElytraItem;
+
+@Mixin(CapeFeatureRenderer.class)
+public class CapeFeatureRendererMixin {
+
+    @Inject(
+            method = "render(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/client/render/VertexConsumerProvider;ILnet/minecraft/client/network/AbstractClientPlayerEntity;FFFFFF)V",
+            at = @At("HEAD"),
+            cancellable = true
+    )
+    private void miapi$customItemRendering(MatrixStack matrixStack, VertexConsumerProvider vertexConsumerProvider, int i, AbstractClientPlayerEntity abstractClientPlayerEntity, float f, float g, float h, float j, float k, float l, CallbackInfo ci){
+        ItemStack itemStack = abstractClientPlayerEntity.getEquippedStack(EquipmentSlot.CHEST);
+        if (itemStack.getItem() instanceof ModularElytraItem) {
+            ci.cancel();
+        }
+    }
+}